Ajustando videoUrl e função de tirar like e favoritar

+++ src/controllers/favoritesController.ts
@@ -34,9 +34,13 @@ export const favoritesController = {
   },
   delete: async (req: AuthenticatedRequest, res: Response) => {
     const userId = req.user!.id;
-    const courseId = req.params.id;
+    const courseId = req.params.courseId;
 
     try {
+      if (!courseId) {
+        return res.status(400).json({ message: "Course ID é obrigatório." });
+      }
+
       await favoriteService.delete(userId, Number(courseId));
       return res.status(204).send();
     } catch (err) {
