+++ src/controllers/likesController.ts
@@ -17,8 +17,13 @@ export const likesController = {
   },
   delete: async (req: AuthenticatedRequest, res: Response) => {
     const userId = req.user!.id;
-    const { courseId } = req.body;
+    const courseId = req.params.courseId;
+
     try {
+      if (!courseId) {
+        return res.status(400).json({ message: "Course ID é obrigatório." });
+      }
+
       await likeService.delete(userId, Number(courseId));
       return res.status(204).send();
     } catch (err) {
